+++ StepSequencerManager.py
@@ -128,7 +128,10 @@ class StepSequencerManager:
             playhead = int(self.song.current_song_time * self._resolution) % 32
             grid_data[playhead] = 2
 
-        message = SysExEncoder.encode_step_sequencer_state(grid_data)
+        message = SysExEncoder.encode_step_sequencer_state(
+            grid_data,
+            page=self._page
+        )
         if message:
             self.c_surface._send_midi(tuple(message))